chore: rename handler files to match function names
