+++ packages/storage-azure/src/deleteFile.ts
