+++ packages/storage-gcs/src/adapter.ts
@@ -5,10 +5,10 @@ import type {
   GeneratedAdapter,
 } from '@payloadcms/plugin-cloud-storage/types'
 
+import { deleteFile } from './deleteFile.js'
 import { generateURL } from './generateURL.js'
-import { deleteFile } from './handleDelete.js'
-import { uploadFile } from './handleUpload.js'
-import { getFile } from './staticHandler.js'
+import { getFile } from './getFile.js'
+import { uploadFile } from './uploadFile.js'
 
 interface CreateGcsAdapterArgs {
   acl?: 'Private' | 'Public'
