+++ packages/storage-gcs/src/deleteFile.ts
