+++ packages/storage-r2/src/deleteFile.ts
