+++ packages/storage-s3/src/adapter.ts
@@ -5,12 +5,12 @@ import type {
   GeneratedAdapter,
 } from '@payloadcms/plugin-cloud-storage/types'
 
-import type { SignedDownloadsConfig } from './staticHandler.js'
+import type { SignedDownloadsConfig } from './getFile.js'
 
+import { deleteFile } from './deleteFile.js'
 import { generateURL } from './generateURL.js'
-import { deleteFile } from './handleDelete.js'
-import { uploadFile } from './handleUpload.js'
-import { getFile } from './staticHandler.js'
+import { getFile } from './getFile.js'
+import { uploadFile } from './uploadFile.js'
 
 interface CreateS3AdapterArgs {
   acl?: 'private' | 'public-read'
