+++ packages/storage-uploadthing/src/adapter.ts
@@ -8,10 +8,10 @@ import type { UTApi } from 'uploadthing/server'
 
 import type { ACL } from './index.js'
 
+import { deleteFile } from './deleteFile.js'
 import { generateURL } from './generateURL.js'
-import { deleteFile } from './handleDelete.js'
-import { uploadFile } from './handleUpload.js'
-import { getFile } from './staticHandler.js'
+import { getFile } from './getFile.js'
+import { uploadFile } from './uploadFile.js'
 
 interface CreateUploadthingAdapterArgs {
   acl: ACL
